fix: fix bug when update and delete sosmed accont when user is invalid owner

+++ src/api/controllers/SosmedAccount/deleteSosmedAccountById.js
@@ -1,6 +1,8 @@
 import { request, response } from "express";
 import sosmedAccountService from "../../../lib/services/SosmedAccount";
 import NotFoundError from "../../../utils/exceptions/NotFoundError";
+import AuthorizationError from "../../../utils/exceptions/AuthorizationError";
+import tokenize from "../../../utils/tokenize";
 
 const deleteSosmedAccountById = async (req = request, res = response) => {
   const { id } = req.params;
@@ -11,6 +13,19 @@ const deleteSosmedAccountById = async (req = request, res = response) => {
 
   if (!sosmedAccount) {
     throw new NotFoundError("Sosmed account not found, put valid id");
+  }
+
+  const token = req.headers["authorization"];
+  const { id: user_id } = await tokenize.decodeJWT(token);
+
+  const checkOwnerSosmedAccount =
+    await sosmedAccountService.checkOwnerSosmedAccount(
+      sosmedAccount.id,
+      user_id
+    );
+
+  if (!checkOwnerSosmedAccount) {
+    throw new AuthorizationError("User not authorized to access");
   }
 
   await sosmedAccountService.deleteSosmedAccountById(Number(id));
